backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
openai-whisper==20231117
ffmpeg-python==0.2.0 